grOSS-main/requirements.txt
asgiref==3.5.2
attrs==21.4.0
bandit==1.7.4
beautifulsoup4==4.11.1
boltons==21.0.0
bracex==2.3.post1
bs4==0.0.1
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-option-group==0.5.3
colorama==0.4.5
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.5
djangorestframework==3.13.1
face==20.1.1
gitdb==4.0.9
GitPython==3.1.27
gittyleaks==0.0.31
glom==22.1.0
idna==3.3
jschema-to-python==1.2.3
jsonpickle==2.2.0
jsonschema==4.6.0
libsast==1.5.2
lxml==4.9.1
mobsfscan==0.1.1
njsscan==0.3.1
packaging==21.3
pbr==5.9.0
peewee==3.14.10
pycparser==2.21
PyGithub==1.55
PyJWT==2.4.0
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-lsp-jsonrpc==1.0.0
pytz==2022.1
PyYAML==6.0
requests==2.28.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sarif-om==1.0.4
scandir==1.10.0
semgrep==0.104.0
sh==1.14.2
smmap==5.0.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stevedore==3.5.0
tabulate==0.8.10
tqdm==4.64.0
typing_extensions==4.3.0
ujson==5.4.0
urllib3==1.26.9
wcmatch==8.3
wrapt==1.14.1
